+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/Constants.kt
@@ -27,6 +27,7 @@ object Constants {
         const val SETTINGS = "${Plugin.ID}.settings"
 
         object External {
+            const val CHANGELOG = "org.jetbrains.changelog"
             const val IDEA = "idea"
             const val JAVA_TEST_FIXTURES = "java-test-fixtures"
             const val KOTLIN = "org.jetbrains.kotlin.jvm"